chore: split DevServerManifest.swift into one file per type
DevServerSurface, DevServerLocation, DevServerCandidates, and
DevServerLocator move to their own files; DevServerManifest keeps the
manifest model. No behavior change.
